# improve performance / allocations on CreateObject / ResolveArgument